refactor: Extract useDatasetCount
